+++ src/Store.ts
@@ -1,7 +1,7 @@
 import { compose, createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 
-import { rootReducer } from './reducers';
+import { rootReducer } from './modules';
 
 const composeEnhancers = compose;
